Adds python3 support
## Context
Python 3.12 deprecated the `unicode` method in favor of `str`. As a result, systems running Python >= 3.12 throw `NameError: name 'unicode' is not defined` when running this plugin.

## Fix
This PR adds some code to detected whether the `unicode` method is available. If it isn't, it substitutes `str`, which enables the rest of the code to work as-is.

Note that although I've tested this on a system with Python >= 3.12 (I'm running `3.9.12`), I haven't confirmed that it still works with systems running Python 2, although I can't imagine why it wouldn't.